Refactor: 중복 함수, 코드 utils로 제어
