+++ diary/src/utils/emotionImgList.js
@@ -1,0 +1,29 @@
+const emotionList = [
+  {
+    emotionId: 1,
+    emotionImg: `${process.env.PUBLIC_URL}/assets/emotion/emotion1.png`,
+    emotionDescription: '완전 좋음',
+  },
+  {
+    emotionId: 2,
+    emotionImg: `${process.env.PUBLIC_URL}/assets/emotion/emotion2.png`,
+    emotionDescription: '좋음',
+  },
+  {
+    emotionId: 3,
+    emotionImg: `${process.env.PUBLIC_URL}/assets/emotion/emotion3.png`,
+    emotionDescription: '보통',
+  },
+  {
+    emotionId: 4,
+    emotionImg: `${process.env.PUBLIC_URL}/assets/emotion/emotion4.png`,
+    emotionDescription: '나쁨',
+  },
+  {
+    emotionId: 5,
+    emotionImg: `${process.env.PUBLIC_URL}/assets/emotion/emotion5.png`,
+    emotionDescription: '완전 나쁨',
+  },
+];
+
+export default emotionList;
